Enhanced polyglot support
MidiPascal could be called from other GraalVM languages, but there was
no way to call other GraalVM languages from MidiPascal.

For addressing this issue the keyword `POLYGLOT` was introduced.
This keyword allows the specification of the target language and the
command to execute as parameters. It allows parameterizing commands
by using string concatenation introduced earlier.