Update EBNF to Prof. Dobler's wishes
Previously `:=` was uses to separate a non-terminal symbol and its definition.

Prof. Dobler gave some feedback regarding the usage of `:=` and line breaks.
The new convention uses `=` to separate a non-terminal symbol from its definition
and `.` is uses to mark the end of a definition.
Moreover, a line break is inserted following the name of the non-terminal symbol
and the `=` character. These measures should increase readability.